Consolidate netapi tests and move them to reflect naming convention
