--- tests/unit/netapi/rest_cherrypy/__init__.py
@@ -1,1 +1,0 @@
-# -*- coding: utf-8 -*-
