+++ tests/unit/netapi/test_rest_cherrypy.py
